Move big nested classes to another file
