+++ src/Microsoft.Windows.CsWin32/Generator.cs
@@ -24,7 +24,7 @@ namespace Microsoft.Windows.CsWin32;
 /// The core of the source generator.
 /// </summary>
 [DebuggerDisplay("{" + nameof(DebuggerDisplayString) + ",nq}")]
-public class Generator : IDisposable
+public partial class Generator : IDisposable
 {
     internal const string InteropDecorationNamespace = "Windows.Win32.Interop";
     internal const string NativeArrayInfoAttribute = "NativeArrayInfoAttribute";
@@ -6439,443 +6439,6 @@ public class Generator : IDisposable
         internal int? CountConst { get; init; }
     }
 
-    private class GeneratedCode
-    {
-        private readonly GeneratedCode? parent;
-
-        private readonly Dictionary<string, List<MemberDeclarationSyntax>> modulesAndMembers = new Dictionary<string, List<MemberDeclarationSyntax>>(StringComparer.OrdinalIgnoreCase);
-
-        /// <summary>
-        /// The structs, enums, delegates and other supporting types for extern methods.
-        /// </summary>
-        private readonly Dictionary<(TypeDefinitionHandle Type, bool HasUnmanagedName), MemberDeclarationSyntax> types = new();
-
-        private readonly Dictionary<FieldDefinitionHandle, (FieldDeclarationSyntax FieldDeclaration, TypeDefinitionHandle? FieldType)> fieldsToSyntax = new();
-
-        private readonly List<ClassDeclarationSyntax> safeHandleTypes = new();
-
-        private readonly Dictionary<string, (MemberDeclarationSyntax Type, bool TopLevel)> specialTypes = new(StringComparer.Ordinal);
-
-        private readonly Dictionary<string, MethodDeclarationSyntax> macros = new(StringComparer.Ordinal);
-
-        /// <summary>
-        /// The set of types that are or have been generated so we don't stack overflow for self-referencing types.
-        /// </summary>
-        private readonly Dictionary<(TypeDefinitionHandle, bool), Exception?> typesGenerating = new();
-
-        /// <summary>
-        /// The set of methods that are or have been generated.
-        /// </summary>
-        private readonly Dictionary<MethodDefinitionHandle, Exception?> methodsGenerating = new();
-
-        /// <summary>
-        /// A collection of the names of special types we are or have generated.
-        /// </summary>
-        private readonly Dictionary<string, Exception?> specialTypesGenerating = new(StringComparer.Ordinal);
-
-        private readonly Dictionary<(string Namespace, string Name), MemberDeclarationSyntax> inlineArrays = new();
-
-        private readonly Dictionary<string, TypeSyntax?> releaseMethodsWithSafeHandleTypesGenerating = new();
-
-        private readonly List<MethodDeclarationSyntax> inlineArrayIndexerExtensionsMembers = new();
-
-        private readonly List<ClassDeclarationSyntax> comInterfaceFriendlyExtensionsMembers = new();
-
-        private bool generating;
-
-        internal GeneratedCode()
-        {
-        }
-
-        internal GeneratedCode(GeneratedCode parent)
-        {
-            this.parent = parent;
-        }
-
-        internal bool IsEmpty => this.modulesAndMembers.Count == 0 && this.types.Count == 0 && this.fieldsToSyntax.Count == 0 && this.safeHandleTypes.Count == 0 && this.specialTypes.Count == 0
-            && this.inlineArrayIndexerExtensionsMembers.Count == 0 && this.comInterfaceFriendlyExtensionsMembers.Count == 0 && this.macros.Count == 0 && this.inlineArrays.Count == 0;
-
-        internal IEnumerable<MemberDeclarationSyntax> GeneratedTypes => this.GetTypesWithInjectedFields()
-            .Concat(this.specialTypes.Values.Where(st => !st.TopLevel).Select(st => st.Type))
-            .Concat(this.safeHandleTypes)
-            .Concat(this.inlineArrays.Values);
-
-        internal IEnumerable<MemberDeclarationSyntax> GeneratedTopLevelTypes => this.specialTypes.Values.Where(st => st.TopLevel).Select(st => st.Type);
-
-        internal IReadOnlyCollection<ClassDeclarationSyntax> ComInterfaceExtensions => this.comInterfaceFriendlyExtensionsMembers;
-
-        internal IEnumerable<MethodDeclarationSyntax> InlineArrayIndexerExtensions => this.inlineArrayIndexerExtensionsMembers;
-
-        internal IEnumerable<FieldDeclarationSyntax> TopLevelFields => from field in this.fieldsToSyntax.Values
-                                                                       where field.FieldType is null || !this.types.ContainsKey((field.FieldType.Value, false))
-                                                                       select field.FieldDeclaration;
-
-        internal IEnumerable<IGrouping<string, MemberDeclarationSyntax>> MembersByModule
-        {
-            get
-            {
-                foreach (KeyValuePair<string, List<MemberDeclarationSyntax>> item in this.modulesAndMembers)
-                {
-                    yield return new Grouping<string, MemberDeclarationSyntax>(item.Key, item.Value);
-                }
-            }
-        }
-
-        internal IEnumerable<MethodDeclarationSyntax> Macros => this.macros.Values;
-
-        internal void AddSafeHandleType(ClassDeclarationSyntax safeHandleDeclaration)
-        {
-            this.ThrowIfNotGenerating();
-
-            this.safeHandleTypes.Add(safeHandleDeclaration);
-        }
-
-        internal void AddMemberToModule(string moduleName, MemberDeclarationSyntax member)
-        {
-            this.ThrowIfNotGenerating();
-
-            if (!this.modulesAndMembers.TryGetValue(moduleName, out List<MemberDeclarationSyntax>? methodsList))
-            {
-                this.modulesAndMembers.Add(moduleName, methodsList = new List<MemberDeclarationSyntax>());
-            }
-
-            methodsList.Add(member);
-        }
-
-        internal void AddMemberToModule(string moduleName, IEnumerable<MemberDeclarationSyntax> members)
-        {
-            this.ThrowIfNotGenerating();
-
-            if (!this.modulesAndMembers.TryGetValue(moduleName, out List<MemberDeclarationSyntax>? methodsList))
-            {
-                this.modulesAndMembers.Add(moduleName, methodsList = new List<MemberDeclarationSyntax>());
-            }
-
-            methodsList.AddRange(members);
-        }
-
-        internal void AddConstant(FieldDefinitionHandle fieldDefHandle, FieldDeclarationSyntax constantDeclaration, TypeDefinitionHandle? fieldType)
-        {
-            this.ThrowIfNotGenerating();
-            this.fieldsToSyntax.Add(fieldDefHandle, (constantDeclaration, fieldType));
-        }
-
-        internal void AddMacro(string macroName, MethodDeclarationSyntax macro)
-        {
-            this.ThrowIfNotGenerating();
-            this.macros.Add(macroName, macro);
-        }
-
-        internal void AddInlineArrayIndexerExtension(MethodDeclarationSyntax inlineIndexer)
-        {
-            this.ThrowIfNotGenerating();
-
-            string thisParameter = inlineIndexer.ParameterList.Parameters[0].Type!.ToString();
-
-            IEnumerable<MethodDeclarationSyntax> toSearch = this.inlineArrayIndexerExtensionsMembers;
-            if (this.parent is not null)
-            {
-                toSearch = toSearch.Concat(this.parent.inlineArrayIndexerExtensionsMembers);
-            }
-
-            if (!toSearch.Any(m => m.Identifier.ValueText == inlineIndexer.Identifier.ValueText && m.ParameterList.Parameters[0].Type!.ToString() == thisParameter))
-            {
-                this.inlineArrayIndexerExtensionsMembers.Add(inlineIndexer);
-            }
-        }
-
-        internal void AddComInterfaceExtension(ClassDeclarationSyntax extension)
-        {
-            this.ThrowIfNotGenerating();
-            this.comInterfaceFriendlyExtensionsMembers.Add(extension);
-        }
-
-        internal void AddComInterfaceExtension(IEnumerable<ClassDeclarationSyntax> extension)
-        {
-            this.ThrowIfNotGenerating();
-            this.comInterfaceFriendlyExtensionsMembers.AddRange(extension);
-        }
-
-        /// <summary>
-        /// Adds a declaration to the generated code.
-        /// </summary>
-        /// <param name="specialName">The same constant provided to <see cref="GenerateSpecialType(string, Action)"/>. This serves to avoid repeat declarations.</param>
-        /// <param name="specialDeclaration">The declaration.</param>
-        /// <param name="topLevel"><see langword="true" /> if this declaration should <em>not</em> be nested within the top-level namespace for generated code.</param>
-        internal void AddSpecialType(string specialName, MemberDeclarationSyntax specialDeclaration, bool topLevel = false)
-        {
-            this.ThrowIfNotGenerating();
-            this.specialTypes.Add(specialName, (specialDeclaration, topLevel));
-        }
-
-        internal void AddInteropType(TypeDefinitionHandle typeDefinitionHandle, bool hasUnmanagedName, MemberDeclarationSyntax typeDeclaration)
-        {
-            this.ThrowIfNotGenerating();
-            this.types.Add((typeDefinitionHandle, hasUnmanagedName), typeDeclaration);
-        }
-
-        internal void GenerationTransaction(Action generator)
-        {
-            if (this.parent is null)
-            {
-                throw new InvalidOperationException("Code generation should occur in a volatile instance.");
-            }
-
-            if (this.generating)
-            {
-                // A transaction is already running. Just run the generator.
-                generator();
-                return;
-            }
-
-            try
-            {
-                this.generating = true;
-                generator();
-                this.Commit(this.parent);
-            }
-            catch
-            {
-                this.Commit(null);
-                throw;
-            }
-            finally
-            {
-                this.generating = false;
-            }
-        }
-
-        internal void GenerateMethod(MethodDefinitionHandle methodDefinitionHandle, Action generator)
-        {
-            this.ThrowIfNotGenerating();
-
-            if (this.methodsGenerating.TryGetValue(methodDefinitionHandle, out Exception? failure) || this.parent?.methodsGenerating.TryGetValue(methodDefinitionHandle, out failure) is true)
-            {
-                if (failure is object)
-                {
-                    throw new GenerationFailedException("This member already failed in generation previously.", failure);
-                }
-
-                return;
-            }
-
-            this.methodsGenerating.Add(methodDefinitionHandle, null);
-            try
-            {
-                generator();
-            }
-            catch (Exception ex)
-            {
-                this.methodsGenerating[methodDefinitionHandle] = ex;
-                throw;
-            }
-        }
-
-        internal void GenerateSpecialType(string name, Action generator)
-        {
-            this.ThrowIfNotGenerating();
-
-            if (this.specialTypesGenerating.TryGetValue(name, out Exception? failure) || this.parent?.specialTypesGenerating.TryGetValue(name, out failure) is true)
-            {
-                if (failure is object)
-                {
-                    throw new GenerationFailedException("This type already failed in generation previously.", failure);
-                }
-
-                return;
-            }
-
-            this.specialTypesGenerating.Add(name, null);
-            try
-            {
-                generator();
-            }
-            catch (Exception ex)
-            {
-                this.specialTypesGenerating[name] = ex;
-                throw;
-            }
-        }
-
-        internal bool IsInlineArrayStructGenerated(string @namespace, string name) => this.parent?.inlineArrays.ContainsKey((@namespace, name)) is true || this.inlineArrays.ContainsKey((@namespace, name));
-
-        internal void AddInlineArrayStruct(string @namespace, string name, MemberDeclarationSyntax inlineArrayStructDeclaration)
-        {
-            this.ThrowIfNotGenerating();
-
-            this.inlineArrays.Add((@namespace, name), inlineArrayStructDeclaration);
-        }
-
-        internal void GenerateType(TypeDefinitionHandle typeDefinitionHandle, bool hasUnmanagedName, Action generator)
-        {
-            this.ThrowIfNotGenerating();
-
-            var key = (typeDefinitionHandle, hasUnmanagedName);
-            if (this.typesGenerating.TryGetValue(key, out Exception? failure) || this.parent?.typesGenerating.TryGetValue(key, out failure) is true)
-            {
-                if (failure is object)
-                {
-                    throw new GenerationFailedException("This type already failed in generation previously.", failure);
-                }
-
-                return;
-            }
-
-            this.typesGenerating.Add(key, null);
-            try
-            {
-                generator();
-            }
-            catch (Exception ex)
-            {
-                this.typesGenerating[key] = ex;
-                throw;
-            }
-        }
-
-        internal void GenerateConstant(FieldDefinitionHandle fieldDefinitionHandle, Action generator)
-        {
-            this.ThrowIfNotGenerating();
-
-            if (this.fieldsToSyntax.ContainsKey(fieldDefinitionHandle) || this.parent?.fieldsToSyntax.ContainsKey(fieldDefinitionHandle) is true)
-            {
-                return;
-            }
-
-            generator();
-        }
-
-        internal void GenerateMacro(string macroName, Action generator)
-        {
-            this.ThrowIfNotGenerating();
-
-            if (this.macros.ContainsKey(macroName) || this.parent?.macros.ContainsKey(macroName) is true)
-            {
-                return;
-            }
-
-            generator();
-        }
-
-        internal bool TryGetSafeHandleForReleaseMethod(string releaseMethod, out TypeSyntax? safeHandleType)
-        {
-            return this.releaseMethodsWithSafeHandleTypesGenerating.TryGetValue(releaseMethod, out safeHandleType)
-                || this.parent?.releaseMethodsWithSafeHandleTypesGenerating.TryGetValue(releaseMethod, out safeHandleType) is true;
-        }
-
-        internal void AddSafeHandleNameForReleaseMethod(string releaseMethod, TypeSyntax? safeHandleType)
-        {
-            this.ThrowIfNotGenerating();
-
-            this.releaseMethodsWithSafeHandleTypesGenerating.Add(releaseMethod, safeHandleType);
-        }
-
-        private static void Commit<TKey, TValue>(Dictionary<TKey, TValue> source, Dictionary<TKey, TValue>? target)
-        {
-            if (target is object)
-            {
-                foreach (KeyValuePair<TKey, TValue> item in source)
-                {
-                    target.Add(item.Key, item.Value);
-                }
-            }
-
-            source.Clear();
-        }
-
-        private static void Commit<T>(List<T> source, List<T>? target)
-        {
-            if (target is object)
-            {
-                target.AddRange(source);
-            }
-
-            source.Clear();
-        }
-
-        private void Commit(GeneratedCode? parent)
-        {
-            foreach (KeyValuePair<string, List<MemberDeclarationSyntax>> item in this.modulesAndMembers)
-            {
-                if (parent is object)
-                {
-                    if (!parent.modulesAndMembers.TryGetValue(item.Key, out List<MemberDeclarationSyntax>? list))
-                    {
-                        parent.modulesAndMembers.Add(item.Key, list = new());
-                    }
-
-                    list.AddRange(item.Value);
-                }
-
-                item.Value.Clear();
-            }
-
-            Commit(this.types, parent?.types);
-            Commit(this.fieldsToSyntax, parent?.fieldsToSyntax);
-            Commit(this.safeHandleTypes, parent?.safeHandleTypes);
-            Commit(this.specialTypes, parent?.specialTypes);
-            Commit(this.typesGenerating, parent?.typesGenerating);
-            Commit(this.macros, parent?.macros);
-            Commit(this.methodsGenerating, parent?.methodsGenerating);
-            Commit(this.specialTypesGenerating, parent?.specialTypesGenerating);
-            Commit(this.inlineArrays, parent?.inlineArrays);
-            Commit(this.releaseMethodsWithSafeHandleTypesGenerating, parent?.releaseMethodsWithSafeHandleTypesGenerating);
-            Commit(this.inlineArrayIndexerExtensionsMembers, parent?.inlineArrayIndexerExtensionsMembers);
-            Commit(this.comInterfaceFriendlyExtensionsMembers, parent?.comInterfaceFriendlyExtensionsMembers);
-        }
-
-        private IEnumerable<MemberDeclarationSyntax> GetTypesWithInjectedFields()
-        {
-            var fieldsByType =
-                (from field in this.fieldsToSyntax
-                 where field.Value.FieldType is not null
-                 group field.Value.FieldDeclaration by field.Value.FieldType into typeGroup
-                 select typeGroup).ToDictionary(k => k.Key!, k => k.ToArray());
-            foreach (var pair in this.types)
-            {
-                MemberDeclarationSyntax type = pair.Value;
-                if (fieldsByType.TryGetValue(pair.Key.Type, out var extraFields))
-                {
-                    switch (type)
-                    {
-                        case StructDeclarationSyntax structType:
-                            type = structType.AddMembers(extraFields);
-                            break;
-                        default:
-                            throw new NotSupportedException();
-                    }
-                }
-
-                yield return type;
-            }
-        }
-
-        private void ThrowIfNotGenerating()
-        {
-            if (!this.generating)
-            {
-                throw new InvalidOperationException("Generating code must take place within a recognized top-level call.");
-            }
-        }
-
-        private class Grouping<TKey, TElement> : IGrouping<TKey, TElement>
-        {
-            private readonly IEnumerable<TElement> values;
-
-            internal Grouping(TKey key, IEnumerable<TElement> values)
-            {
-                this.Key = key;
-                this.values = values;
-            }
-
-            public TKey Key { get; }
-
-            public IEnumerator<TElement> GetEnumerator() => this.values.GetEnumerator();
-
-            System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => this.GetEnumerator();
-        }
-    }
-
     private class DirectiveTriviaRemover : CSharpSyntaxRewriter
     {
         internal static readonly DirectiveTriviaRemover Instance = new();
@@ -6890,407 +6453,5 @@ public class Generator : IDisposable
             trivia.IsKind(SyntaxKind.EndIfDirectiveTrivia) ||
             trivia.IsKind(SyntaxKind.DisabledTextTrivia)
             ? default : trivia;
-    }
-
-    private class WhitespaceRewriter : CSharpSyntaxRewriter
-    {
-        private readonly List<SyntaxTrivia> indentationLevels = new List<SyntaxTrivia> { default };
-        private int indentationLevel;
-
-        internal WhitespaceRewriter()
-            : base(visitIntoStructuredTrivia: true)
-        {
-        }
-
-        private SyntaxTrivia IndentTrivia => this.indentationLevels[this.indentationLevel];
-
-        private SyntaxTrivia OuterIndentTrivia => this.indentationLevels[Math.Max(0, this.indentationLevel - 1)];
-
-        public override SyntaxNode? VisitNamespaceDeclaration(NamespaceDeclarationSyntax node)
-        {
-            node = node
-                .WithNamespaceKeyword(node.NamespaceKeyword.WithLeadingTrivia(TriviaList(this.IndentTrivia)))
-                .WithOpenBraceToken(node.OpenBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)))
-                .WithCloseBraceToken(node.CloseBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)));
-            using var indent = new Indent(this);
-            SyntaxNode? result = base.VisitNamespaceDeclaration(node);
-            if (result is NamespaceDeclarationSyntax ns)
-            {
-                result = ns.WithMembers(AddSpacingBetweenMembers(ns.Members, ns.Usings.Count > 0));
-            }
-
-            return result;
-        }
-
-        public override SyntaxNode? VisitStructDeclaration(StructDeclarationSyntax node)
-        {
-            node = this.WithIndentingTrivia(node)
-                .WithOpenBraceToken(node.OpenBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)))
-                .WithCloseBraceToken(node.CloseBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)));
-            using var indent = new Indent(this);
-            SyntaxNode? result = base.VisitStructDeclaration(node);
-            if (result is StructDeclarationSyntax s)
-            {
-                result = s.WithMembers(AddSpacingBetweenMembers(s.Members));
-            }
-
-            return result;
-        }
-
-        public override SyntaxNode? VisitClassDeclaration(ClassDeclarationSyntax node)
-        {
-            node = this.WithIndentingTrivia(node)
-                .WithOpenBraceToken(node.OpenBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)))
-                .WithCloseBraceToken(node.CloseBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)));
-            using var indent = new Indent(this);
-            SyntaxNode? result = base.VisitClassDeclaration(node);
-            if (result is ClassDeclarationSyntax c)
-            {
-                result = c.WithMembers(AddSpacingBetweenMembers(c.Members));
-            }
-
-            return result;
-        }
-
-        public override SyntaxNode? VisitInterfaceDeclaration(InterfaceDeclarationSyntax node)
-        {
-            node = this.WithIndentingTrivia(node)
-                .WithOpenBraceToken(node.OpenBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)))
-                .WithCloseBraceToken(node.CloseBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)));
-            using var indent = new Indent(this);
-            SyntaxNode? result = base.VisitInterfaceDeclaration(node);
-            if (result is InterfaceDeclarationSyntax c)
-            {
-                result = c.WithMembers(AddSpacingBetweenMembers(c.Members));
-            }
-
-            return result;
-        }
-
-        public override SyntaxNode? VisitEnumDeclaration(EnumDeclarationSyntax node)
-        {
-            node = this.WithIndentingTrivia(node)
-                .WithOpenBraceToken(node.OpenBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)))
-                .WithCloseBraceToken(node.CloseBraceToken.WithLeadingTrivia(TriviaList(this.IndentTrivia)));
-            using var indent = new Indent(this);
-            return base.VisitEnumDeclaration(node);
-        }
-
-        public override SyntaxNode? VisitUsingDirective(UsingDirectiveSyntax node)
-        {
-            return base.VisitUsingDirective(node.WithLeadingTrivia(this.IndentTrivia));
-        }
-
-        public override SyntaxNode? VisitBlock(BlockSyntax node)
-        {
-            SyntaxTriviaList leadingTrivia;
-            if (node.Parent is FixedStatementSyntax or AccessorDeclarationSyntax or TryStatementSyntax or FinallyClauseSyntax)
-            {
-                leadingTrivia = TriviaList(this.IndentTrivia);
-            }
-            else
-            {
-                leadingTrivia = TriviaList(LineFeed).Add(this.IndentTrivia);
-            }
-
-            node = node
-                .WithOpenBraceToken(Token(leadingTrivia, SyntaxKind.OpenBraceToken, TriviaList(LineFeed)))
-                .WithCloseBraceToken(Token(TriviaList(this.IndentTrivia), SyntaxKind.CloseBraceToken, TriviaList(LineFeed)));
-            using var indent = new Indent(this);
-            return base.VisitBlock(node);
-        }
-
-        public override SyntaxNode? VisitBaseList(BaseListSyntax node)
-        {
-            if (node.Parent is EnumDeclarationSyntax)
-            {
-                return base.VisitBaseList(node);
-            }
-            else
-            {
-                return base.VisitBaseList(this.WithIndentingTrivia(node));
-            }
-        }
-
-        public override SyntaxNode? VisitAttributeList(AttributeListSyntax node)
-        {
-            if (node.Parent is ParameterSyntax)
-            {
-                return node.WithCloseBracketToken(TokenWithSpace(SyntaxKind.CloseBracketToken));
-            }
-            else if (node.Parent is BaseTypeDeclarationSyntax)
-            {
-                return this.WithOuterIndentingTrivia(node);
-            }
-            else
-            {
-                return this.WithIndentingTrivia(node);
-            }
-        }
-
-        public override SyntaxNode? VisitMethodDeclaration(MethodDeclarationSyntax node) => base.VisitMethodDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitConstructorDeclaration(ConstructorDeclarationSyntax node) => base.VisitConstructorDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitOperatorDeclaration(OperatorDeclarationSyntax node) => base.VisitOperatorDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitConversionOperatorDeclaration(ConversionOperatorDeclarationSyntax node) => base.VisitConversionOperatorDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitDelegateDeclaration(DelegateDeclarationSyntax node) => base.VisitDelegateDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitFieldDeclaration(FieldDeclarationSyntax node) => base.VisitFieldDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitEnumMemberDeclaration(EnumMemberDeclarationSyntax node) => base.VisitEnumMemberDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitPropertyDeclaration(PropertyDeclarationSyntax node) => base.VisitPropertyDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitIndexerDeclaration(IndexerDeclarationSyntax node) => base.VisitIndexerDeclaration(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitAccessorList(AccessorListSyntax node)
-        {
-            node = node
-                .WithOpenBraceToken(Token(TriviaList(this.IndentTrivia), SyntaxKind.OpenBraceToken, TriviaList(LineFeed)))
-                .WithCloseBraceToken(Token(TriviaList(this.IndentTrivia), SyntaxKind.CloseBraceToken, TriviaList(LineFeed)));
-            using var indent = new Indent(this);
-            return base.VisitAccessorList(node);
-        }
-
-        public override SyntaxNode? VisitAccessorDeclaration(AccessorDeclarationSyntax node)
-        {
-            if (node.Body is not null)
-            {
-                node = node.WithKeyword(node.Keyword.WithTrailingTrivia(LineFeed));
-            }
-
-            return base.VisitAccessorDeclaration(this.WithIndentingTrivia(node));
-        }
-
-        public override SyntaxNode? VisitLocalDeclarationStatement(LocalDeclarationStatementSyntax node) => base.VisitLocalDeclarationStatement(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitExpressionStatement(ExpressionStatementSyntax node) => base.VisitExpressionStatement(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitInitializerExpression(InitializerExpressionSyntax node)
-        {
-            node = node
-                .WithOpenBraceToken(Token(TriviaList(this.IndentTrivia), SyntaxKind.OpenBraceToken, TriviaList(LineFeed)))
-                .WithCloseBraceToken(Token(TriviaList(this.IndentTrivia), SyntaxKind.CloseBraceToken, TriviaList(LineFeed)))
-                .WithTrailingTrivia(TriviaList());
-            using var indent = new Indent(this);
-            return base.VisitInitializerExpression(node);
-        }
-
-        public override SyntaxNode? VisitAssignmentExpression(AssignmentExpressionSyntax node)
-        {
-            if (node.Parent is InitializerExpressionSyntax)
-            {
-                return base.VisitAssignmentExpression(this.WithIndentingTrivia(node));
-            }
-            else
-            {
-                return base.VisitAssignmentExpression(node);
-            }
-        }
-
-        public override SyntaxNode? VisitTryStatement(TryStatementSyntax node) => base.VisitTryStatement(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitFinallyClause(FinallyClauseSyntax node) => base.VisitFinallyClause(this.WithIndentingTrivia(node));
-
-        public override SyntaxNode? VisitIfStatement(IfStatementSyntax node)
-        {
-            node = this.WithIndentingTrivia(node);
-            if (node.Statement is BlockSyntax)
-            {
-                return base.VisitIfStatement(node);
-            }
-            else
-            {
-                using var indent = new Indent(this);
-                return base.VisitIfStatement(node);
-            }
-        }
-
-        public override SyntaxNode? VisitWhileStatement(WhileStatementSyntax node)
-        {
-            node = this.WithIndentingTrivia(node);
-            if (node.Statement is BlockSyntax)
-            {
-                return base.VisitWhileStatement(node);
-            }
-            else
-            {
-                using var indent = new Indent(this);
-                return base.VisitWhileStatement(node);
-            }
-        }
-
-        public override SyntaxNode? VisitElseClause(ElseClauseSyntax node)
-        {
-            node = this.WithIndentingTrivia(node);
-            if (node.Statement is BlockSyntax)
-            {
-                return base.VisitElseClause(node);
-            }
-            else
-            {
-                using var indent = new Indent(this);
-                return base.VisitElseClause(node);
-            }
-        }
-
-        public override SyntaxNode? VisitFixedStatement(FixedStatementSyntax node)
-        {
-            node = this.WithIndentingTrivia(node);
-            if (node.Statement is BlockSyntax)
-            {
-                return base.VisitFixedStatement(node);
-            }
-            else
-            {
-                using var indent = new Indent(this);
-                return base.VisitFixedStatement(node);
-            }
-        }
-
-        public override SyntaxNode? VisitForStatement(ForStatementSyntax node)
-        {
-            node = this.WithIndentingTrivia(node);
-            if (node.Statement is BlockSyntax)
-            {
-                return base.VisitForStatement(node);
-            }
-            else
-            {
-                using var indent = new Indent(this);
-                return base.VisitForStatement(node);
-            }
-        }
-
-        public override SyntaxNode? VisitReturnStatement(ReturnStatementSyntax node)
-        {
-            return base.VisitReturnStatement(this.WithIndentingTrivia(node));
-        }
-
-        public override SyntaxToken VisitToken(SyntaxToken token)
-        {
-            if (token.IsKind(SyntaxKind.CommaToken) && token.Parent is ParameterListSyntax or AttributeArgumentListSyntax or ArgumentListSyntax)
-            {
-                return TokenWithSpace(SyntaxKind.CommaToken);
-            }
-
-            return base.VisitToken(token);
-        }
-
-        public override SyntaxTriviaList VisitList(SyntaxTriviaList list)
-        {
-#if DEBUG && false // Nodes that contain any annotations at all cause a lot of lock contention that slows us down. Consider removing it all and enforcing (part of it) with this code
-            if (list.Any() && list[0].IsEquivalentTo(SyntaxFactory.ElasticMarker))
-            {
-                throw new GenerationFailedException("Elastic trivia got by us.");
-            }
-#endif
-
-            string? indent = null;
-            for (int i = list.Count - 1; i >= 0; i--)
-            {
-                if (list[i].GetStructure() is DocumentationCommentTriviaSyntax trivia)
-                {
-                    indent ??= list[i].Token.Parent is BaseTypeDeclarationSyntax or AttributeListSyntax { Parent: BaseTypeDeclarationSyntax } ? this.OuterIndentTrivia.ToString() : this.IndentTrivia.ToString();
-                    var comment = new StringBuilder(trivia.Content.ToFullString());
-                    comment.Insert(0, indent);
-                    comment.Replace("\n", "\n" + indent);
-                    comment.Length -= indent.Length; // Remove the extra indent after the last newline.
-                    list = list.RemoveAt(i).InsertRange(i, ParseLeadingTrivia(comment.ToString()));
-                }
-            }
-
-            return list; // do not recurse into trivia
-        }
-
-        private static SyntaxList<MemberDeclarationSyntax> AddSpacingBetweenMembers(SyntaxList<MemberDeclarationSyntax> members, bool insertLineAboveFirstMember = false)
-        {
-            List<MemberDeclarationSyntax> mutableMembers = members.ToList();
-            for (int i = mutableMembers.Count - 1; i > 0; i--)
-            {
-                if (mutableMembers[i] is
-                    ClassDeclarationSyntax or
-                    StructDeclarationSyntax or
-                    NamespaceDeclarationSyntax or
-                    EnumDeclarationSyntax or
-                    MethodDeclarationSyntax or
-                    IndexerDeclarationSyntax or
-                    PropertyDeclarationSyntax)
-                {
-                    mutableMembers[i] = mutableMembers[i].WithLeadingTrivia(mutableMembers[i].GetLeadingTrivia().Insert(0, LineFeed));
-                }
-            }
-
-            if (insertLineAboveFirstMember && mutableMembers.Count > 0)
-            {
-                mutableMembers[0] = mutableMembers[0].WithLeadingTrivia(mutableMembers[0].GetLeadingTrivia().Insert(0, LineFeed));
-            }
-
-            return new SyntaxList<MemberDeclarationSyntax>(mutableMembers);
-        }
-
-        private TSyntax WithIndentingTrivia<TSyntax>(TSyntax node, SyntaxTrivia indentTrivia)
-            where TSyntax : SyntaxNode
-        {
-            if (node is MemberDeclarationSyntax memberDeclaration)
-            {
-                SyntaxToken firstToken = GetFirstToken(memberDeclaration);
-                return node.ReplaceToken(firstToken, firstToken.WithLeadingTrivia(firstToken.HasLeadingTrivia ? firstToken.LeadingTrivia.Add(indentTrivia) : TriviaList(indentTrivia)));
-            }
-
-            // Take care to preserve xml doc comments, pragmas, etc.
-            return node.WithLeadingTrivia(node.HasLeadingTrivia ? this.VisitList(node.GetLeadingTrivia()).Add(indentTrivia) : TriviaList(indentTrivia));
-
-            static SyntaxToken GetFirstToken(MemberDeclarationSyntax memberDeclaration)
-            {
-                if (!memberDeclaration.AttributeLists.Any())
-                {
-                    return memberDeclaration.GetFirstToken();
-                }
-                else if (memberDeclaration.Modifiers.Any())
-                {
-                    return memberDeclaration.Modifiers[0];
-                }
-                else
-                {
-                    return memberDeclaration.GetFirstToken();
-                }
-            }
-        }
-
-        private TSyntax WithIndentingTrivia<TSyntax>(TSyntax node)
-            where TSyntax : SyntaxNode
-        {
-            return this.WithIndentingTrivia(node, this.IndentTrivia);
-        }
-
-        private TSyntax WithOuterIndentingTrivia<TSyntax>(TSyntax node)
-            where TSyntax : SyntaxNode
-        {
-            return this.WithIndentingTrivia(node, this.OuterIndentTrivia);
-        }
-
-        private struct Indent : IDisposable
-        {
-            private readonly WhitespaceRewriter rewriter;
-
-            internal Indent(WhitespaceRewriter rewriter)
-            {
-                this.rewriter = rewriter;
-                rewriter.indentationLevel++;
-                for (int i = rewriter.indentationLevels.Count; i <= rewriter.indentationLevel; i++)
-                {
-                    rewriter.indentationLevels.Add(SyntaxTrivia(SyntaxKind.WhitespaceTrivia, new string('\t', i)));
-                }
-            }
-
-            public void Dispose()
-            {
-                this.rewriter.indentationLevel--;
-            }
-        }
     }
 }
